- revert some constants to app specific alias imports - remove currencyCodesOfCountry from helper - fix locale values in localization service - move types to models
