--- src/consts/games.ts
@@ -1,36 +1,0 @@
-export enum SlugCategoriesGames {
-    SLUG_CATEGORY_NEW = "new",
-    SLUG_CATEGORY_BONUS_WAGERING = "bonus-wagering",
-    SLUG_CATEGORY_MEGAWAYS = "megaways",
-    SLUG_CATEGORY_BONUS_BUY = "bonus-buy",
-    SLUG_CATEGORY_BLACKJACK = "blackjack",
-    SLUG_CATEGORY_ROULETTE = "roulette",
-    SLUG_CATEGORY_BACCARAT = "baccarat",
-    SLUG_CATEGORY_POKER = "poker",
-    SLUG_CATEGORY_TABLE = "table",
-    SLUG_CATEGORY_BTC_GAMES = "btc-games", // cryptoGame
-    SLUG_CATEGORY_GAME_SHOWS = "game-shows",
-    SLUG_CATEGORY_MYSTIC_JACKPOTS = "mystic-jackpots",
-    SLUG_CATEGORY_BNG_TOURNAMENT = "bng-tournament",
-    SLUG_CATEGORY_JACKPOT = "jackpot",
-    SLUG_CATEGORY_PLS_TOURNAMENT = "pls-tournament",
-    SLUG_CATEGORY_TOP = "top",
-    SLUG_CATEGORY_RECOMMENDED = "recommended",
-    SLUG_CATEGORY_EGYPT_GAMES = "egypt-games",
-    SLUG_CATEGORY_FRUITS_GAMES = "fruit-games",
-    SLUG_CATEGORY_BOOKS_GAMES = "book-games",
-    SLUG_CATEGORY_CRASH_GAMES = "crash-games",
-    SLUG_CATEGORY_LOTTERY_GAMES = "lottery-games",
-    SLUG_CATEGORY_FISHING_GAMES = "fishing-games",
-    SLUG_CATEGORY_HIGH_VOLATILITY = "high-volatility",
-    SLUG_CATEGORY_LOW_VOLATILITY = "low-volatility",
-    SLUG_CATEGORY_SPECIAL = "special",
-    SLUG_CATEGORY_TOUR_GAMES = "tour_games",
-    SLUG_CATEGORY_ALL = "all",
-    SLUG_CATEGORY_LIVE = "live",
-    SLUG_CATEGORY_POKIES = "pokies",
-    SLUG_CATEGORY_WINN_WEEK = "winning-week",
-    SLUG_CATEGORY_WIN_WEEK_TOUR = "winning-week-tournament",
-    SLUG_CATEGORY_COINS = "coin",
-    SLUG_CATEGORY_LIVE_ALL = "live_all",
-}
