+++ src/models/enums/currencies.ts
@@ -21,4 +21,19 @@ export enum Currencies {
     TRX = "TRX",
 }
 
+export const currencyCodesOfCountry = {
+    AU: "AUD",
+    CA: "CAD",
+    NZ: "NZD",
+    US: "USD",
+    IE: "EUR",
+    DE: "EUR",
+    JP: "JPY",
+    NO: "NOK",
+    BR: "BRL",
+    AT: "EUR",
+    CH: "EUR",
+};
+
+
 export default {};
