+++ src/services/user.ts
@@ -1,3 +1,5 @@
+import { ID_GROUP_FOR_PAIRED_ID, ID_GROUP_FOR_UNPAIRED_ID } from "@config/groupAB";
+
 import { useUserInfo } from "../store/user/userInfo";
 import { useUserStatuses } from "../store/user/userStatuses";
 import { addPlayerToGroup } from "./api/requests/player";
@@ -6,8 +8,6 @@ export async function userSetToGroupForAbTest() {
     const userInfo = useUserInfo();
     const userStatuses = useUserStatuses();
 
-    const ID_GROUP_FOR_PAIRED_ID = 543;
-    const ID_GROUP_FOR_UNPAIRED_ID = 544;
     const isUserIncludingInAB = userStatuses.getUserGroups.some((id) => {
         return id === ID_GROUP_FOR_PAIRED_ID || id === ID_GROUP_FOR_UNPAIRED_ID;
     });
